requirements.txt
py-cord==2.7.0
pydub==0.25.1
python-dotenv==1.2.1
